Expense Creation Logic for Fix Price and Percentage

+++ db/migrate/20241205210047_remove_user_id_from_products.rb
@@ -1,5 +1,7 @@
 class RemoveUserIdFromProducts < ActiveRecord::Migration[7.1]
   def change
-    remove_column :products, :user_id, :bigint
+    if column_exists?(:products, :user_id)
+      remove_column :products, :user_id, :bigint
+    end
   end
 end
